removal of jnativehook.jar since we are getting an actual button board, reorganization of auto modes/ states/ statemachines and removal of depricated states/autos/statemachines
