+++ src/main/java/frc/team555/robot/core/PowerUpRobot.java
@@ -3,14 +3,16 @@ package frc.team555.robot.core;
 import com.ctre.phoenix.motorcontrol.can.WPI_TalonSRX;
 import edu.wpi.first.wpilibj.CameraServer;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
-import frc.team555.robot.automodes.*;
-import frc.team555.robot.states.*;
+import frc.team555.robot.stateMachines.AutoClimbSequence;
+import frc.team555.robot.stateMachines.CubeOuttake;
+import frc.team555.robot.stateMachines.SwitchAuto;
+import frc.team555.robot.stateMachines.automodes.*;
+import frc.team555.robot.stateMachines.states.*;
 import frc.team555.robot.components.CubeIntake;
 import frc.team555.robot.components.IntakeLift;
 import frc.team555.robot.components.MainLift;
 import frc.team555.robot.utils.CoastMotor;
 import org.montclairrobotics.sprocket.SprocketRobot;
-import org.montclairrobotics.sprocket.auto.AutoMode;
 import org.montclairrobotics.sprocket.auto.states.*;
 import org.montclairrobotics.sprocket.control.*;
 import org.montclairrobotics.sprocket.drive.*;
