+++ src/main/java/frc/team555/robot/stateMachines/AutoEndGame.java
@@ -1,7 +1,6 @@
-package frc.team555.robot.states;
+package frc.team555.robot.stateMachines;
 
 import frc.team555.robot.components.MainLift;
-import frc.team555.robot.core.PowerUpRobot;
 import org.montclairrobotics.sprocket.auto.states.DriveEncoderGyro;
 import org.montclairrobotics.sprocket.geometry.*;
 import org.montclairrobotics.sprocket.states.StateMachine;
