+++ src/main/java/frc/team555/robot/stateMachines/DropCube.java
@@ -1,5 +1,7 @@
-package frc.team555.robot.states;
+package frc.team555.robot.stateMachines;
 
+import frc.team555.robot.stateMachines.states.MoveLift;
+import frc.team555.robot.stateMachines.states.SideTurn;
 import frc.team555.robot.components.CubeIntake;
 import frc.team555.robot.components.MainLift;
 import frc.team555.robot.utils.Side;
