+++ src/main/java/frc/team555/robot/stateMachines/TopCubeAuto.java
@@ -1,5 +1,6 @@
-package frc.team555.robot.states;
+package frc.team555.robot.stateMachines;
 
+import frc.team555.robot.stateMachines.states.ConditionalState;
 import frc.team555.robot.components.CubeIntake;
 import frc.team555.robot.components.MainLift;
 import frc.team555.robot.utils.Side;
