+++ src/main/java/frc/team555/robot/stateMachines/automodes/CenterBaseLineRight.java
@@ -1,4 +1,4 @@
-package frc.team555.robot.automodes;
+package frc.team555.robot.stateMachines.automodes;
 
 import org.montclairrobotics.sprocket.auto.AutoMode;
 import org.montclairrobotics.sprocket.auto.states.DriveEncoderGyro;
@@ -7,13 +7,13 @@ import org.montclairrobotics.sprocket.geometry.Degrees;
 
 import static frc.team555.robot.core.PowerUpRobot.correction;
 
-public class CenterBaseLineLeft extends AutoMode {
+public class CenterBaseLineRight extends AutoMode{
 
-    public CenterBaseLineLeft() {
-        super("Center Base Line Left",
+    public CenterBaseLineRight(){
+        super("Center Base Line Right",
                 new ResetGyro(correction),
                 new DriveEncoderGyro(60, .5, new Degrees(0), false, correction),
-                new DriveEncoderGyro(12, .5, new Degrees(-90), false, correction),
+                new DriveEncoderGyro(122, .5, new Degrees(90), false, correction),
                 new DriveEncoderGyro(80, .5, new Degrees(0), false, correction)
         );
     }
