+++ src/main/java/frc/team555/robot/stateMachines/states/ConditionalState.java
@@ -1,4 +1,4 @@
-package frc.team555.robot.states;
+package frc.team555.robot.stateMachines.states;
 
 import org.montclairrobotics.sprocket.states.State;
 import org.montclairrobotics.sprocket.utils.Input;
