--- src/main/java/frc/team555/robot/states/LiftState.java
@@ -1,23 +1,0 @@
-package frc.team555.robot.states;
-
-import frc.team555.robot.components.IntakeLift;
-import org.montclairrobotics.sprocket.auto.AutoState;
-@Deprecated
-public class LiftState extends AutoState {
-
-    IntakeLift intakeLift;
-    int position;
-
-    public LiftState(int position){
-        this.position = position;
-    }
-    @Override
-    public void stateUpdate() {
-        intakeLift.setPosition(position);
-        isDone();
-
-    }
-
-    @Override
-    public boolean isDone() { return true; }
-}
